+++ pages/api/roles/index.ts
@@ -3,8 +3,8 @@ import type { NextApiRequest, NextApiResponse } from 'next';
 import { RoleData, RoleListData, RolePostData } from '../../../types/role';
 import { withAnyUser } from '../../../utils/auth/jwtHelpers';
 import { HTTP_GET_METHOD, HTTP_POST_METHOD, HTTP_STATUS_CREATED, HTTP_STATUS_OK } from '../../../utils/http/httpHelper';
+import { rejectHttpMethodsNotIn } from '../../../utils/http/rejectHttpMethodsNotIn';
 import { createIfPossible } from '../../../utils/prisma/createIfPossible';
-import { rejectHttpMethodsNotIn } from '../../../utils/prisma/rejectHttpMethodsNotIn';
 
 const prisma = new PrismaClient();
